s/readers_cache: fixed disposing reader when gate is closed
An error in `storage::readers_cache` implementation may lead to the
situation in which closing the log might stuck and never return. The
issue was caused by the fact that background reader disposal was not
executed when `readers_cache` was closed. This way the not reusable
reader was never deleted and hold the `log` lock lease preventing log
from being closed. Fixed an issue by making sure that reader will be
gracefully closed and disposed during `readers_cache` shutdown.

Signed-off-by: Michal Maslanka <michal@vectorized.io>